requirements.txt
Django==4.2.3
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
requests==2.31.0
gunicorn
django-cors-headers==4.2.0

